Reorganize tests, closes #69
